Scan numbers
Support reading in numbers in the scanner.
Some more refactoring. Changed the scanner bit from
Option<Result<TokenType, ScannerError> to just Result<TokenType,
ScannerError> since a None was just mapping to a
ScannerError::UnrecognizedToken anyway.

This also cleaned up an early return for unterminated strings that was
kind of ugly.